[feat] #85 : QueryDSL - response dto 개발
기본적으로 판매점마다 1등, 2등 당첨 횟수를 집계할 필요가 있는데, 이러한 부분을 포함해 querydsl 적용 범위를 확장한다.

기존에 사용하던 response dto 클래스도 정상적으로 동작했었다.
querydsl 적용을 위한 `projection class` 개발 및 불필요한 dto 는 삭제한다.
